python/requirements.txt
azure-eventhub
azure-identity
aiohttp
azure-eventhub-checkpointstoreblob-aio
